behavioural - mediator - problem
Button should be enabled if listbox or text box have value. Other wise it should be disabled. User should not be able to click that button.
Othercase, if use choose something in ListBox, the choosed filename should be updated in text box.
Like, user choosed test.txt file in listbox and text box automatically updated with the value text.txt.

Here we should not couple ListBox, TextBox and Button directly since it's a framework details. It may be used different cases in different places.
mediator pattern solves this problem